Moves Dataset.get_entities to __init__ module.
Addresses sixth part of #477.
